requirements.txt
h5py>=3.13.0
jupyter>=1.1.1
matplotlib>=3.10.1
numpy>=2.2.4
protobuf>=6.30.2
pycollada>=0.9
pynput>=1.8.1
pyyaml>=6.0.2
pyzmq>=26.3.0
rerun-sdk>=0.22.1
scipy>=1.15.2
trimesh>=4.6.6
ur_rtde==1.6.0
urdf_parser_py>=0.0.4
